+++ tools/mailer/mailer.go
@@ -9,15 +9,16 @@ import (
 
 // Message defines a generic email message struct.
 type Message struct {
-	From        mail.Address         `json:"from"`
-	To          []mail.Address       `json:"to"`
-	Bcc         []mail.Address       `json:"bcc"`
-	Cc          []mail.Address       `json:"cc"`
-	Subject     string               `json:"subject"`
-	HTML        string               `json:"html"`
-	Text        string               `json:"text"`
-	Headers     map[string]string    `json:"headers"`
-	Attachments map[string]io.Reader `json:"attachments"`
+	From              mail.Address         `json:"from"`
+	To                []mail.Address       `json:"to"`
+	Bcc               []mail.Address       `json:"bcc"`
+	Cc                []mail.Address       `json:"cc"`
+	Subject           string               `json:"subject"`
+	HTML              string               `json:"html"`
+	Text              string               `json:"text"`
+	Headers           map[string]string    `json:"headers"`
+	Attachments       map[string]io.Reader `json:"attachments"`
+	InlineAttachments map[string]io.Reader `json:"inlineAttachments"`
 }
 
 // Mailer defines a base mail client interface.